third_party/WebKit/LayoutTests/svg/animations/animate-path-animation-qQ-tT-inverse-expected.txt

SVG 1.1 dynamic animation tests
Test path animation where coordinate modes of start and end differ. You should see PASS messages

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS path.getAttribute('d') is "M -30 -30 q 30 0 30 30 t -30 30 Z"
PASS path.getAttribute('d') is "M -30 -30 q 37.5 0 37.5 30 t -37.5 30 Z"
PASS path.getAttribute('d') is "M -30 -30 Q 22.5 -30 22.5 0 T -30 30 Z"
PASS path.getAttribute('d') is "M -30 -30 Q 29.9925 -30 29.9925 0 T -30 30 Z"
PASS path.getAttribute('d') is "M -30 -30 q 30 0 30 30 t -30 30 Z"
PASS successfullyParsed is true

TEST COMPLETE
